+++ Collier/Host/SignalRBackgroundService.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Collier.Hubs;
-using Collier.Mining.OutputParsing;
+using Collier.Mining;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
