+++ Collier/Hubs/ICollierClient.cs
@@ -5,17 +5,13 @@ using System.Threading.Tasks;
 
 namespace Collier.Hubs
 {
+    /// <summary>
+    /// Represents what methods we can call on the client from the server.  I ended up doing most
+    /// event notification generically in SignalRBackgroundService so this interface didn't get built
+    /// out as much as planned.  
+    /// </summary>
     public interface ICollierClient
     {
-        Task UpdateMinerLogMessage(string message);
-        Task UpdateHashRate(string newRate);
-        Task UpdateTemperature(string temp);
-        Task UpdateCrashInfo(string temp);
-
-        Task UpdatePower(string power);
-
-        Task UpdateState(string state);
-
         //SignalR clients do not have a listener handler for connection events.
         //you can chain off the start connection event, but that's cubmersome if we
         //want objects farther down the chain to act on connection.
