+++ Collier/Mining/IMiner.cs
@@ -1,4 +1,6 @@
 ﻿using Collier.Mining.State;
+using Collier.Mining.Trex;
+using Collier.Mining.Trex.State;
 using System.Threading.Tasks;
 
 namespace Collier.Mining
