+++ Collier/Mining/IMiningInfoNotifier.cs
@@ -3,8 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Collier.Mining.OutputParsing
+namespace Collier.Mining
 {
+    /// <summary>
+    /// Any type of subscribable event around mining that should be pushed to clients
+    /// </summary>
     public interface IMiningInfoNotifier
     {
         event EventHandler<MiningInformation> MiningInformationChanged;
