+++ Collier/Mining/LogMessage.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Collier.Mining.OutputParsing
+namespace Collier.Mining
 {
+    //a single line of output from the log file
     public class LogMessage
     {
         public string Message { get; set; }
