+++ CollierTests/Mining/MinerEventStateTests.cs
@@ -7,8 +7,8 @@ using System.Threading.Tasks;
 using Collier.Host;
 using Collier.IO;
 using Collier.Mining;
-using Collier.Mining.OutputParsing;
 using Collier.Mining.State;
+using Collier.Mining.Trex;
 using FluentAssertions;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
